Dev2

+++ Models/OfertaEmpleo.cs
@@ -13,7 +13,7 @@ namespace SisEmpleo.Models
         public string titulo { get; set; }
         public string descripcion { get; set; }
         public int vacante { get; set; }
-        public float salario { get; set; }
+        public double salario { get; set; }
         public string horario { get; set; }
         public string duracion_contrato { get; set; }
         public DateTime fecha_publicacion { get; set; }
